modified todo of section04_step02
