+++ hands-on/gacha/sec4/gacha/card.go
@@ -1,4 +1,4 @@
-package main
+package gacha
 
 type rarity string
 
@@ -10,16 +10,16 @@ rarityXR rarity = "XR"
 )
 
 func (r rarity) String() string {
-return string(r)
+	return string(r)
 }
 
 type card struct {
-rarity rarity // レア度
-name   string // 名前
+	rarity rarity // レア度
+	name   string // 名前
 }
 
 func (c *card) String() string {
-// TODO: レア度:名前のように文字列を作る
-// 例："SR:ドラゴン"
+
+
 return c.rarity.String() + ":" + c.name
 }
